services/blog/requirements.txt
alembic==1.8.1
amqp==5.1.1
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
boto3==1.26.5
botocore==1.29.5
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.1
flasgger==0.9.5
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.0
gunicorn==20.1.0
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.16.0
kombu==5.2.4
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
mistune==2.0.4
packaging==21.3
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.33
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.4
python-logstash==0.4.8
pytz==2022.6
PyYAML==6.0
redis==4.3.5
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.42
SQLAlchemy-Utils==0.38.3
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.10.0
